fix babelify
