+++ babelify/babelify-tests.ts
@@ -1,5 +1,3 @@
-/// <reference path="babelify.d.ts" />
-
 import babelify = require("babelify");
 
 module BabelifyTest {
